fix: cast result to double.

+++ lib/pages/Teacher/Classes/AstaGraphs.dart
@@ -502,7 +502,8 @@ class _AstaGraphsState extends State<AstaGraphs> {
                                                     builder: (context) =>
                                                         AstaSubGraph(
                                                       passedSummResult:
-                                                          dt['Result'],
+                                                          dt['Result']
+                                                              .toDouble(),
                                                       passedSummDate:
                                                           dt['Created']
                                                               as Timestamp,
